pkg/lakego-pkg/go-cryptobin/cipher/square/square_test.go
package square

import (
    "bytes"
    "testing"
    "math/rand"
)

func Test_Square(t *testing.T) {
    random := rand.New(rand.NewSource(99))
    max := 5000

    var encrypted [32]byte
    var decrypted [32]byte

    for i := 0; i < max; i++ {
        key := make([]byte, 32)
        random.Read(key)
        value := make([]byte, 32)
        random.Read(value)

        cipher1, err := NewCipher(key)
        if err != nil {
            t.Fatal(err.Error())
        }

        cipher1.XORKeyStream(encrypted[:], value)

        cipher2, err := NewCipher(key)
        if err != nil {
            t.Fatal(err.Error())
        }

        cipher2.XORKeyStream(decrypted[:], encrypted[:])

        if !bytes.Equal(decrypted[:], value[:]) {
            t.Errorf("encryption/decryption failed: % 02x != % 02x\n", decrypted, value)
        }
    }
}
